+++ docs/components/mdxComponents.tsx
@@ -1,4 +1,3 @@
-// TODO Rename this file `mdxComponents`
 import { AnchorHTMLAttributes, HTMLAttributes, Fragment } from 'react';
 import Link from 'next/link';
 import { slugify } from '../lib/slugify';
