+++ docs/pages/guides/[slug].tsx
@@ -8,7 +8,7 @@ import {
 	getGuideSlugs,
 	Guide,
 } from '../../lib/mdx/guides';
-import { mdxComponents } from '../../components/utils';
+import { mdxComponents } from '../../components/mdxComponents';
 import { AppLayout } from '../../components/AppLayout';
 import { DocumentTitle } from '../../components/DocumentTitle';
 import { PageLayout } from '../../components/PageLayout';
