+++ docs/pages/releases/[slug].tsx
@@ -9,7 +9,7 @@ import {
 	getReleaseSlugs,
 	Release,
 } from '../../lib/mdxUtils';
-import { mdxComponents } from '../../components/utils';
+import { mdxComponents } from '../../components/mdxComponents';
 import { AppLayout } from '../../components/AppLayout';
 import { DocumentTitle } from '../../components/DocumentTitle';
 import { PageLayout } from '../../components/PageLayout';
