+++ docs/pages/templates/index.tsx
@@ -8,7 +8,7 @@ import { mq } from '@ag.ds-next/core';
 import { Text } from '@ag.ds-next/text';
 import { getMarkdownData, serializeMarkdown } from '../../lib/mdxUtils';
 import { getTemplateList } from '../../lib/mdx/templates';
-import { mdxComponents } from '../../components/utils';
+import { mdxComponents } from '../../components/mdxComponents';
 import { AppLayout } from '../../components/AppLayout';
 import { PageLayout } from '../../components/PageLayout';
 import { DocumentTitle } from '../../components/DocumentTitle';
